Adding ability to directly set display data
Although the underlying structure of the driver's panel data is stored in the 7 bit frame format the panels receive, the method takes in fully saturated bytestrings representing screen state. As a user of the library, it makes more sense like this and you shouldn't worry about the underlying structures.

Also adds an overridden version of drawText that doesn't refresh the screen. This helps eliminate unnecessary refreshes that would slow down fps.